flaskapp/requirements.txt
flask
redis
Flask-SQLAlchemy
flask-restful
mysqlclient
pymysql
cerberus
gunicorn
flask-bootstrap